+++ test/pipeline_test.exs
@@ -23,6 +23,11 @@ defmodule PipelineTest do
       send(self(), {:hook, state})
     end
 
+    def error_handler(_state, _options) do
+      send(self(), :error)
+      :error
+    end
+
     def single_async_hook(state, options) do
       send(options[:parent_pid], {:async_hook, state})
     end
@@ -45,6 +50,11 @@ defmodule PipelineTest do
       send(self(), {:hook, state})
     end
 
+    def error_handler(%Pipeline.State{error: error}, _options) do
+      send(self(), {:handler, error})
+      error
+    end
+
     def single_async_hook(state, options) do
       send(options[:parent_pid], {:async_hook, state})
     end
@@ -59,8 +69,8 @@ defmodule PipelineTest do
       assert function_exported?(__MODULE__.GoodPipeline, :execute, 2)
     end
 
-    test "the injected __pipeline__ returns only the matched functions for steps, hooks and async hooks" do
-      {steps, hooks, async_hooks} = __MODULE__.GoodPipeline.__pipeline__()
+    test "the injected __pipeline__ returns only the matched functions for steps, handlers, hooks and async hooks" do
+      {steps, hooks, async_hooks, handlers} = __MODULE__.GoodPipeline.__pipeline__()
 
       assert steps == [
                {__MODULE__.GoodPipeline, :first_step},
@@ -74,6 +84,10 @@ defmodule PipelineTest do
       assert async_hooks == [
                {__MODULE__.GoodPipeline, :single_async_hook}
              ]
+
+      assert handlers == [
+               {__MODULE__.GoodPipeline, :error_handler}
+             ]
     end
   end
 
@@ -83,6 +97,7 @@ defmodule PipelineTest do
       assert result == {:ok, 12}
       assert_receive {:first, 10}
       assert_receive {:second, 11}
+      refute_receive {:handler, :error}
 
       assert_receive {:hook,
                       %Pipeline.State{valid?: true, initial_value: 10, value: 12, error: nil}}
@@ -91,12 +106,14 @@ defmodule PipelineTest do
                       %Pipeline.State{valid?: true, initial_value: 10, value: 12, error: nil}}
     end
 
-    test "do not execute steps after a failure, but still call hooks anyways" do
+    test "do not execute steps after a failure, but still call handlers and hooks anyways" do
       result = Pipeline.execute(__MODULE__.PipelineWithError, 0, parent_pid: self())
       assert result == {:error, "Some error"}
 
       assert_receive {:first, 0}
       refute_receive {:second, 1}
+
+      assert_receive {:handler, "Some error"}
 
       assert_receive {:hook,
                       %Pipeline.State{